Rename `IconBase` as `FontAwesomeIcon`
This is with the aim of allowing the generic Icon component to be used
with the dedicated className passed in as prop.
