+++ src/components/icons/FontAwesomeIcon/FontAwesomeIcon.d.ts
@@ -1,5 +1,5 @@
-import DefaultMobile from './IconBase.mobile';
-import * as mobile from './IconBase.mobile';
+import DefaultMobile from './FontAwesomeIcon.mobile';
+import * as mobile from './FontAwesomeIcon.mobile';
 // import DefaultWeb from './AddTerm.web';
 // import * as web from './AddTerm.web';
 
@@ -9,5 +9,5 @@ declare var _test: typeof mobile;
 declare var _testDefault: typeof DefaultMobile;
 // declare var _testDefault: typeof DefaultWeb;
 
-export * from './IconBase.mobile';
+export * from './FontAwesomeIcon.mobile';
 export default DefaultMobile;
