+++ src/components/icons/FontAwesomeIcon/FontAwesomeIcon.mobile.tsx
