+++ src/components/icons/PencilIcon/PencilIcon.mobile.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import styled, { StyledFunction } from 'styled-components';
 
-import IconBase from '@src/components/icons/IconBase/IconBase'
+import FontAwesomeIcon from '@src/components/icons/FontAwesomeIcon/FontAwesomeIcon'
 
-const PencilIcon = (props) => <IconBase {...props} icon="fa-pencil"/>;
+const PencilIcon = (props) => <FontAwesomeIcon {...props} icon="fa-pencil"/>;
 
 export default PencilIcon;
