+++ src/components/icons/SearchIcon/SearchIcon.mobile.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import styled, { StyledFunction } from 'styled-components';
 
-import IconBase from '@src/components/icons/IconBase/IconBase'
+import FontAwesomeIcon from '@src/components/icons/FontAwesomeIcon/FontAwesomeIcon'
 
-const SearchIcon = (props) => <IconBase {...props} icon="fa-search"/>;
+const SearchIcon = (props) => <FontAwesomeIcon {...props} icon="fa-search"/>;
 
 export default SearchIcon;
