+++ src/components/icons/ThreeDot/ThreeDot.mobile.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import styled, { StyledFunction } from 'styled-components';
 
 import { Span, I } from '@src/styles/elems'; 
-import IconBase from '@src/components/icons/IconBase/IconBase';
+import FontAwesomeIcon from '@src/components/icons/FontAwesomeIcon/FontAwesomeIcon';
 
-const ThreeDot = () => <IconBase icon="fa-ellipsis-h"/>;
+const ThreeDot = () => <FontAwesomeIcon icon="fa-ellipsis-h"/>;
 
 export default ThreeDot;
