+++ src/components/icons/Upvote/Upvote.mobile.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import styled, { StyledFunction } from 'styled-components';
 
-import IconBase from '../IconBase/IconBase';
+import FontAwesomeIcon from '../FontAwesomeIcon/FontAwesomeIcon';
 
-const Upvote = () => <IconBase icon="fa-thumbs-up"/>;
+const Upvote = () => <FontAwesomeIcon icon="fa-thumbs-up"/>;
 
 export default Upvote;
